+++ src/components/contactDetail/PeopleDetailPanel.tsx
